app/requirements.txt
# Flask og relaterte pakker
flask==2.0.1
werkzeug==2.0.3
sqlalchemy==1.4.46
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
flask-jwt-extended==4.3.1
pymysql==1.0.2
cryptography==3.4.7
python-dotenv==0.19.0
